+++ apps/taiga-lumbermill/src/dashboards/crypto/components/minter/minter-created/minter-created.component.html
@@ -1,13 +1,13 @@
 <tui-avatar
     size="xl"
     class="logo"
-    [src]="tokenMinter.urlIcon"
+    [src]="token.urlIcon"
 />
 <h3
     tuiTitle
     class="info"
 >
-    {{ tokenMinter.amount }} (${{ tokenMinter.symbol }}) have been sent to your wallet
+    {{ token.amount }} (${{ token.symbol }}) have been sent to your wallet
 </h3>
 <button
     tuiButton
